requirements.txt
pyyaml
fastapi
uvicorn
pydantic
peewee
psycopg2-binary
pyjwt
requests
